NetworkOptions does not declare anymore its state (besides activity logging).
Motivation:

TCP related configuration started to be encapsulated into TcpOptions but only the TCPSSLOptions portion was moved to this class.

In order to be fully reusable, the NetworkOptions portion should be declared as well.

Changes:

- NetworkOptions does not declare anymore state, related property accessors have been made abstract
- TcpOptions declares the NetworkOptions state as well as its property accessors
- TCPSSLOptions implements the property accessors defined by NetworkOptions by delegating to TcpOptions
- DatagramSocketOptions declares the NetworkOptions state and implements the property accessors defined by NetworkOptions
