+++ vertx-core/src/main/java/io/vertx/core/datagram/DatagramSocketOptions.java
@@ -57,6 +57,10 @@ public class DatagramSocketOptions extends NetworkOptions {
    */
   public static final boolean DEFAULT_IPV6 = false;
 
+  private int sendBufferSize;
+  private int receiveBufferSize;
+  private int trafficClass;
+  private boolean reuseAddress;
   private boolean reusePort;
   private boolean broadcast;
   private boolean loopbackModeDisabled;
@@ -80,6 +84,10 @@ public class DatagramSocketOptions extends NetworkOptions {
    */
   public DatagramSocketOptions(DatagramSocketOptions other) {
     super(other);
+    this.sendBufferSize = other.getSendBufferSize();
+    this.receiveBufferSize = other.getReceiveBufferSize();
+    this.reuseAddress = other.isReuseAddress();
+    this.trafficClass = other.getTrafficClass();
     this.reusePort = other.reusePort;
     this.broadcast = other.isBroadcast();
     this.loopbackModeDisabled = other.isLoopbackModeDisabled();
@@ -100,6 +108,10 @@ public class DatagramSocketOptions extends NetworkOptions {
   }
 
   private void init() {
+    sendBufferSize = DEFAULT_SEND_BUFFER_SIZE;
+    receiveBufferSize = DEFAULT_RECEIVE_BUFFER_SIZE;
+    reuseAddress = DEFAULT_REUSE_ADDRESS;
+    trafficClass = DEFAULT_TRAFFIC_CLASS;
     reusePort = DEFAULT_REUSE_PORT;
     broadcast = DEFAULT_BROADCAST;
     loopbackModeDisabled = DEFAULT_LOOPBACK_MODE_DISABLED;
@@ -121,40 +133,48 @@ public class DatagramSocketOptions extends NetworkOptions {
 
   @Override
   public int getSendBufferSize() {
-    return super.getSendBufferSize();
+    return sendBufferSize;
   }
 
   @Override
   public DatagramSocketOptions setSendBufferSize(int sendBufferSize) {
-    super.setSendBufferSize(sendBufferSize);
+    Arguments.require(sendBufferSize > 0  || sendBufferSize == DEFAULT_SEND_BUFFER_SIZE, "sendBufferSize must be > 0");
+    this.sendBufferSize = sendBufferSize;
     return this;
   }
 
   @Override
   public int getReceiveBufferSize() {
-    return super.getReceiveBufferSize();
+    return receiveBufferSize;
   }
 
   @Override
   public DatagramSocketOptions setReceiveBufferSize(int receiveBufferSize) {
-    super.setReceiveBufferSize(receiveBufferSize);
+    Arguments.require(receiveBufferSize > 0 || receiveBufferSize == DEFAULT_RECEIVE_BUFFER_SIZE, "receiveBufferSize must be > 0");
+    this.receiveBufferSize = receiveBufferSize;
     return this;
   }
 
   @Override
+  public boolean isReuseAddress() {
+    return reuseAddress;
+  }
+
+  @Override
   public DatagramSocketOptions setReuseAddress(boolean reuseAddress) {
-    super.setReuseAddress(reuseAddress);
+    this.reuseAddress = reuseAddress;
     return this;
   }
 
   @Override
   public int getTrafficClass() {
-    return super.getTrafficClass();
+    return trafficClass;
   }
 
   @Override
   public DatagramSocketOptions setTrafficClass(int trafficClass) {
-    super.setTrafficClass(trafficClass);
+    Arguments.requireInRange(trafficClass, NetworkOptions.DEFAULT_TRAFFIC_CLASS, 255, "trafficClass tc must be 0 <= tc <= 255");
+    this.trafficClass = trafficClass;
     return this;
   }
